mod jpa to jdbc

+++ CommonSecurity/src/main/java/mx/morena/security/servicio/MasterService.java
@@ -6,14 +6,14 @@ public class MasterService {
 	
 	protected final Logger logger = Logger.getLogger(this.getClass());
 
-	protected static final Integer PERFIL_ESTATAL = 1;
-	protected static final Integer PERFIL_FEDERAL = 2;
-	protected static final Integer PERFIL_LOCAL = 3;
-	protected static final Integer PERFIL_MUNICIPAL = 4;
-	protected static final Integer PERFIL_COTS = 5;
-	protected static final Integer PERFIL_SECCION_ELECTORAL = 6;
-	protected static final Integer PERFIL_CRG = 7;
-	protected static final Integer PERFIL_RG = 8;
-	protected static final Integer PERFIL_RC = 9;
+	protected static final Long PERFIL_ESTATAL = 1L;
+	protected static final Long PERFIL_FEDERAL = 2L;
+	protected static final Long PERFIL_LOCAL = 3L;
+	protected static final Long PERFIL_MUNICIPAL = 4L;
+	protected static final Long PERFIL_COTS = 5L;
+	protected static final Long PERFIL_SECCION_ELECTORAL = 6L;
+	protected static final Long PERFIL_CRG = 7L;
+	protected static final Long PERFIL_RG = 8L;
+	protected static final Long PERFIL_RC = 9L;
 
 }
